pyproject.toml
[tool.poetry]
name = "xmas-card-generator"
version = "0.1.0"
description = ""
authors = ["Vladimir Gerasimenko <vladworldss@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.1.0"

[tool.poetry.dev-dependencies]
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.8.0"

[tool.pytest.ini_options]
testpaths = "tests/unit"
python_files = "*.py"
addopts = "-rfs -p no:warnings --tb=short --show-capture=no"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning"
]

[tool.mypy]
allow_redefinition = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
warn_no_return = false
exclude = '^tests/'

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
line_length = 120
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true
lines_between_types = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
